+++ CHANGELOG.md
@@ -25,6 +25,9 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 
 ### Changed
 * Upgraded to Pythia 8.311
+* Rotation infrastructure moved from `DeformedNucleus` class to `Nucleus` class
+* Added a more precise documentation for the orientation section of deformed nuclei and changed the default value for theta
+from pi/2 to 0
 
 ## SMASH-3.1
 Date: 2024-02-26